refactor(ai): Densify Slice D residual fat pins and tighten suite-size gates (Refs #4079)
Extract naval/economy/conquest pin cases and topic-split S7D rollup/lock-recovery
support so residual ≥650 modules fall under tightened soft ceilings (S7D 650;
residual fat pins 650-or-cases).
