+++ packages/colonizethis_ai/test/support/s7d/s7d_campaign_rollup.dart
@@ -1,13 +1,18 @@
-// Per-GP state for the seed-42 S7-D diagnostic campaign (Refs #3997).
+// Per-GP state for the seed-42 S7-D diagnostic campaign (Refs #3997 / #4079 Slice D).
 
 import 'package:colonizethis_ai/colonizethis_ai.dart';
 import 'package:colonizethis_data/colonizethis_data.dart'
     hide cheapestRegimentBuildTreasuryCost;
 
 import 'supply_probes.dart';
+import 's7d_campaign_rollup_explorer_counters.dart';
+import 's7d_campaign_rollup_trade_counters.dart';
 
 /// Owns all mutable observation state for one S7-D campaign run.
-class Seed42S7dCampaignRollup {
+class Seed42S7dCampaignRollup
+    with
+        Seed42S7dCampaignRollupExplorerCounters,
+        Seed42S7dCampaignRollupTradeCounters {
   Seed42S7dCampaignRollup(this.gpIds);
 
   final List<String> gpIds;
@@ -207,93 +212,6 @@ class Seed42S7dCampaignRollup {
     for (final gpId in gpIds)
       gpId: <String, int>{for (final id in castIronFeedstockIds) id: 0},
   };
-
-  // Refs #2847 H8-extraction Old World mineral feedstock prospect
-  // localization (post-#3257 reservation). The reservation holds back an
-  // idle Builder/Explorer for Old World feedstock work, yet
-  // `gpCastIronFeedstockHeldAtTurn99` still shows `iron == 0` for every
-  // supplier (`iron` is never extracted) while surface `timber` is. A
-  // mineral `build_improvement` is rejected until the tile is prospected
-  // (`work_order_target_prechecks.dart`), and only an **idle** Explorer is
-  // reservable, so these two counters split the residual `iron` break,
-  // captured while the supplier castIron gate is active:
-  //
-  //   * `supplierIdleExplorerPresentTurns` — the supplier owns an idle
-  //     Explorer this turn (a unit the reservation could route onto the
-  //     `iron` prospect). A near-zero count localizes the break to
-  //     **Explorer availability** (all Explorers busy / dispatched to
-  //     multi-turn New World exploration, so the reservation never has an
-  //     idle Explorer to hold).
-  //   * `supplierProspectedMineralFeedstockTileTurns` — the supplier owns a
-  //     **prospected** Old World `iron` mineral feedstock tile. A non-zero
-  //     count alongside `iron` held == 0 instead localizes the break
-  //     **downstream** of prospecting (the Builder never improves the
-  //     prospected tile / cannot afford the improvement); a flat zero
-  //     confirms the prospect itself never happens.
-  //   * `supplierIdleExplorerColocatedFeedstockTileTurns` — the supplier
-  //     owns an idle Explorer standing **in the same province** as an
-  //     unprospected Old World `iron` mineral feedstock tile. `prospect`
-  //     candidate generation only reaches an Explorer positioned on (or
-  //     single-hop from) the feedstock province, and the reservation holds
-  //     the lexicographically-smallest idle Explorer **without
-  //     repositioning it**. A flat zero alongside
-  //     `supplierIdleExplorerPresentTurns > 0` localizes the residual to
-  //     reservation **positioning** (no idle Explorer ever reaches the
-  //     feedstock province, so no `prospect` candidate generates); a
-  //     non-zero count instead points at candidate-generation eligibility
-  //     (mineral-tile gate / validator) or selection ranking for an
-  //     already-positioned Explorer.
-  //   * `supplierIdleExplorerColocatedMineralEligibleFeedstockTileTurns` —
-  //     the supplier owns an idle Explorer co-located with an unprospected
-  //     Old World `iron` mineral feedstock tile that **also** passes the
-  //     live mineral-eligibility terrain check (`isMineralEligibleTile`
-  //     under the seed-42 `tileMapByRegion`). This is the next gate the
-  //     `prospect` candidate must clear in
-  //     `_allAcceptedProspectTilesInProvince`. Comparing it against
-  //     `supplierIdleExplorerColocatedFeedstockTileTurns` splits the
-  //     residual finer: a flat zero here while the resource-only co-located
-  //     count is non-zero localizes the break to **terrain
-  //     mineral-eligibility** at candidate generation (the owned `iron`
-  //     tile sits on non-prospectable terrain); equal non-zero counts
-  //     instead point **downstream** of eligibility (validator material
-  //     cost / visibility precheck or selection ranking).
-  //   * `supplierIdleExplorerColocatedSuggestedProspectTileTurns` — the
-  //     **real** `suggestWorkOrders` pass actually emits a `prospect`
-  //     candidate for the co-located mineral-eligible feedstock tile. This
-  //     is the next gate past terrain eligibility: it runs the live
-  //     generation pass (province visibility, move-leg validation, and the
-  //     incremental-validator material-cost / visibility precheck all live
-  //     inside it) rather than re-deriving one gate. Comparing it against
-  //     `supplierIdleExplorerColocatedMineralEligibleFeedstockTileTurns`
-  //     resolves the final fork: a non-zero count proves the prospect is
-  //     generated + validator-accepted, so the residual is **selection
-  //     ranking** (the accepted `prospect` loses to a competing `explore`
-  //     in `selectFullAiCivilianWorkOrders`); a flat zero while the
-  //     mineral-eligible count is non-zero localizes the residual **inside
-  //     generation** (the visibility / move-leg / validator gates), not
-  //     ranking.
-  //
-  // Read-only; the (freely tunable) counts can move as later slices land.
-  late final supplierIdleExplorerPresentTurns = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final supplierProspectedMineralFeedstockTileTurns = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final supplierIdleExplorerColocatedFeedstockTileTurns = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final supplierIdleExplorerColocatedMineralEligibleFeedstockTileTurns =
-      <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final supplierIdleExplorerColocatedSuggestedProspectTileTurns =
-      <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final supplierIdleExplorerColocatedFeedstockProspectProvinceVisibleTurns =
-      <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final supplierIdleExplorerColocatedFeedstockProspectBundledMoveLegTurns =
-      <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final supplierIdleExplorerColocatedFeedstockProspectValidatorTurns =
-      <String, int>{for (final gpId in gpIds) gpId: 0};
-
   // Refs #2847 H8-supply: domestic-production feedstock-stage isolation.
   // The post-#3235 surface shows the world market never supplies fabric
   // (`gpRegimentInputDealsAsBuyer == 0`) and the affluent-supplier release
@@ -643,41 +561,6 @@ class Seed42S7dCampaignRollup {
   late final feedstockAcquisitionTargetWithFieldArmyTurns = <String, int>{
     for (final gpId in gpIds) gpId: 0,
   };
-
-  // Refs #2924 Step 0 — world-market lock-recovery diagnostics:
-  // per-GP rollups capturing (a) trade orders the AI submits each
-  // turn (offer/bid counts plus urgent-priority offer counts at
-  // [kTreasuryOfferPriorityUrgent]), (b) deals matched in the
-  // world-market phase counted by seller/buyer GP plus treasury
-  // credited/debited per side, and (c) whether/when the post-turn
-  // treasury crosses [cheapestRegimentBuildTreasuryCost]. These
-  // surfaces are issue-2924 specific and live alongside the
-  // existing #2847 S7-D fields so a single run produces both
-  // diagnostic blocks.
-  late final tradeOfferCount = <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final tradeUrgentOfferCount = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final tradeBidCount = <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final dealsAsSeller = <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final dealsAsBuyer = <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final treasuryCredited = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final treasuryDebited = <String, int>{for (final gpId in gpIds) gpId: 0};
-  late final regimentThresholdCrossingsUp = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  late final regimentThresholdFirstReachTurn = <String, int?>{
-    for (final gpId in gpIds) gpId: null,
-  };
-  late final treasuryAtTurn99 = <String, int>{
-    for (final gpId in gpIds) gpId: 0,
-  };
-  // Treasury immediately after the previous turn resolved (seeded
-  // from turn-0 pre-resolution treasury so the first crossing
-  // detection compares against game start rather than zero).
-  late final treasuryPrevTurn = <String, int>{};
 
   /// Per-turn scratch state shared between harness callbacks.
   late final pendingTurnScratch = <String, Object>{};
